requirements.txt
streamlit>=1.55.0
python-dotenv>=1.2.2
openai>=2.30.0
azure-core>=1.39.0
azure-search-documents>=11.6.0
azure-ai-documentintelligence>=1.0.2
azure-ai-vision-imageanalysis>=1.0.0
azure-cognitiveservices-speech>=1.48.2
